requirements.txt
boto3==1.9.225
botocore==1.12.225
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
docutils==0.15.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
jmespath==0.9.4
numpy==1.17.2
pandas==0.25.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
